+++ backend/controllers/slackChannels.js
@@ -1,19 +1,8 @@
-const { WebClient, LogLevel } = require('@slack/web-api')
 
-async function slackChannels(slackToken, res){
-
-  const channels = []
-
-  const client = new WebClient(slackToken, {
-    logLevel: LogLevel.DEBUG,
-  })
-
+async function slackChannels(res, slack){
   try {
-    const result = await client.conversations.list({})
-    console.log(result)
-    result.channels.filter(elem => elem.is_channel).forEach(elem => channels.push(elem.name))
-
-    res.send(channels)
+    const result = await slack.getChannelNames()
+    res.send(result)
   } catch (error){
     res.send(error.data.error)
   }
